+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # RELEASE HISTORY
 
+### 0.0.10, / ?, ? 2019
+* fixed password issue when updating the BackendlessUser object
+* fixed Date type issue when retrieving object or registering BackendlessUser with custom Date field
+
 ### 0.0.9, / June, ? 2019
 * added IEmailEnvelope protocol, EmailEnvelope, EnvelopeWithRecepients and EnvelopeWithQuery classes
 * added functions to MessagingService: